+++ services/beaconchain/src/main/java/tech/pegasys/teku/services/beaconchain/BeaconChainController.java
@@ -464,7 +464,7 @@ public class BeaconChainController extends Service
     LOG.debug("BeaconChainController.initForkChoice()");
     final boolean proposerBoostEnabled = beaconConfig.eth2NetworkConfig().isProposerBoostEnabled();
     forkChoice =
-        ForkChoice.create(
+        new ForkChoice(
             spec, forkChoiceExecutor, recentChainData, forkChoiceNotifier, proposerBoostEnabled);
     forkChoiceTrigger = new ForkChoiceTrigger(forkChoice);
   }
@@ -889,6 +889,8 @@ public class BeaconChainController extends Service
     syncService.subscribeToSyncStateChanges(
         syncState -> forkChoiceNotifier.onSyncingStatusChanged(syncState.isInSync()));
     forkChoiceNotifier.onSyncingStatusChanged(syncService.getCurrentSyncState().isInSync());
+
+    forkChoice.subscribeToOptimisticSyncChanges(syncService.getOptimisticSyncSubscriber());
   }
 
   protected void initOperationsReOrgManager() {