+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # Changelog
 
 ## Upcoming Breaking Changes
@@ -20,5 +21,8 @@ For information on changes in released versions of Teku, see the [releases page]
 * Optimised sync committee processing to avoid duplicate group checks for public keys.
 * Reduced amount of data copying required when calculating sha256 hashes and use a more efficient sha256 implementation when available.
 * Updated Javalin to version 4.2.0.
+* Added periodic keep alive to event stream.
 
 ### Bug Fixes
+* Rest api endpoints accepting validator IDs will no longer reject valid bytes48 hex strings that are not on the g2 curve.
+* Upgraded discovery to fix `ConcurrentModificationException`